Atheme Digest Interface: Use an enumeration for the algorithm identifier
This ensures that the compiler will help catch any mistakes with their
given values, instead of relying on runtime code to do so. Specifically:

- Warn if a switch() statement does not handle one of the possible values
  of the enumeration (e.g. code that doesn't handle an algorithm).

- Warn if passing an integer value as the argument to a function when
  that value is not one of the possible values of the enumeration.

+++ modules/crypto/pbkdf2v2.c
@@ -43,7 +43,7 @@ static void
 atheme_pbkdf2v2_config_ready(void ATHEME_VATTR_UNUSED *const restrict unused)
 {
 	if (! pbkdf2v2_digest)
-		pbkdf2v2_digest = PBKDF2_PRF_DEFAULT;
+		pbkdf2v2_digest = (unsigned int) PBKDF2_PRF_DEFAULT;
 
 	if (! pbkdf2v2_rounds)
 		pbkdf2v2_rounds = PBKDF2_ITERCNT_DEF;
